refactor(init): extract flagsToInitOptions into its own file
Move flagsToInitOptions, InitCommandFlags, InitCommandArgs, and the
narrowing helpers out of types.ts into flagsToInitOptions.ts. types.ts
now only contains type definitions.
